requirements.txt
# Telegram Bot Framework
aiogram>=3.4.0

# Morphological Analysis (Russian lemmatization)
pymorphy3>=2.0.0
pymorphy3-dicts-ru>=2.0.0

# Async SQLite
aiosqlite>=0.19.0

# Environment Variables
python-dotenv>=1.0.0
